Add method `trackEvent()` to main exports.
This method is meant to be for event emitters that don't have an instance of
EmitterDetails (cause if it did then you'd just do `emDet.getEventDetails()`).

It's supposed to be a quick way to run an event emitter through the main exports
and register an event to get the instance of EventDetails while having the
listener be of no importance.  Mainly used for mocking an event in a quick
manner.

Fixes #5