lib, test: add getStackTrace()
First, we change the exports of lib/trace.js to be an object.  Then we embellish
the object with the function that used to be the export, `getCallSite()` and
add the new method `getStackTrace()`.

This adds a test and also changes the index.js to use it in one ocassion.

This commit is part of a suite of commits getting the whole shit ready as
mentioned in #8.

Refs: #8