ci: pass build
Note that the way tests are run locally and on a remote build (for CI) differs.  Locally I use test.bat but since [Travis-CI does not have the platform][21] we have to resort to another way and this is it.

Bash throws when it encounters files with parentheses.  For that I've renamed all files with parenthesis to be the same name but without them.

.travis.yml now includes a `script` field that lists all files individually to be run with `node` command.
i.e.
`node <path-to-test-file>`.

I also had to switch to a more cross-platform use of the slash separator.  From now on if we want to build for cross-platform -- and we have to if we're using Travis-CI -- we must use `path.sep`.

[21]: https://github.com/reecehudson/node-emitter-details/issues/21 "GitHub"