ci: fix build failing
Because of some unknown reason -- pretty sure it's cause the command `cd` is
a no-op on our Travis-CI build -- our `process.cwd()` in test/stack-trace.js
was not what is was expected to be (i.e. ./test) and so test/stack-trace.js
was causing the fail.

Now we simply skip that test *only if* we are on Travis-CI.

Refs: #27