lib, test, README: fix fake test of onUpdate()
This correctly tests `EventDetails.prototype.onUpdate()` which was originally
implemented in 2c8f7ef, we correctly do it by making sure we emit the event a
second time so `genericEventRegulator()` is invoked aka updated.  The current
test is replicated and exported to it's own proper test file.

This also corrects docs by saying the only arg passed to the cb.